Change max connections to 1024 for MySQL.
This is because of "Too Many Connections" errors coming from the DB.
In experimentation, were were unable to reproduce this issue. However,
we figure it is likely due to the email bounce handler, and its usage
of DELAYED INSERT statements. Thus, even though the apache process
is complete, the DB connection hangs around until the delayed insert
takes place.

Will restart the mysql server if the file changed as a result of the
update script.